backend/requirements.txt
fastapi>=0.136,<0.137
uvicorn>=0.45,<0.46
python-multipart>=0.0.26,<0.1
sqlalchemy>=2.0,<2.1
opencv-python>=4.11,<5
numpy>=2.0,<3
scikit-image>=0.25,<0.26
scipy>=1.15,<1.16
matplotlib>=3.10,<3.11
pillow>=11.2,<12
httpx>=0.28,<0.29
